Correccion de codigo en la implementacion de imagenes
